refactor: Moves Event API to presentation layer
